Custom html/css/javascript page
My medp150 class only required us to make a html based page with a link (a href onto another page with more text). I decided to go a bit further and incorporate css and javascript. Links to my side project, an old school horror based escape the room game that is currently in the works.